Persian movies Word Cloud is developed.
A Persian font, and coplete csv dataset for movies is added,
Then a .txt dataset of just movie names created and used for Word Cloud.

arabic_reshaper and bidi libraries for Persian Word Cloud are needed and
used. Also Pesian stopwords are used as well.

Telegram or Twitter data will be the next goals!